Extract a contract test for the tenant repository
