+++ Services/CO.CDP.Tenant.Persistence.Tests/DatabaseTenantRepositoryTest.cs
@@ -1,74 +1,14 @@
 using CO.CDP.Testcontainers.PostgreSql;
-using FluentAssertions;
 using Microsoft.EntityFrameworkCore;
-using static CO.CDP.Tenant.Persistence.ITenantRepository.TenantRepositoryException;
 
 namespace CO.CDP.Tenant.Persistence.Tests;
 
-public class DatabaseTenantRepositoryTest(PostgreSqlFixture postgreSql) : IClassFixture<PostgreSqlFixture>
+public class DatabaseTenantRepositoryTest(PostgreSqlFixture postgreSql) :
+    TenantRepositoryContractTest, IClassFixture<PostgreSqlFixture>
 {
-    [Fact]
-    public async Task ItFindsSavedTenant()
+    protected override ITenantRepository TenantRepository()
     {
-        using var repository = new DatabaseTenantRepository(TenantContext());
-
-        var tenant = GivenTenant(guid: Guid.NewGuid());
-
-        repository.Save(tenant);
-
-        var found = await repository.Find(tenant.Guid);
-
-        found.Should().Be(tenant);
-    }
-
-    [Fact]
-    public void ItRejectsTwoTenantsWithTheSameName()
-    {
-        using var repository = new DatabaseTenantRepository(TenantContext());
-
-        var tenant1 = GivenTenant(guid: Guid.NewGuid(), name: "Bob");
-        var tenant2 = GivenTenant(guid: Guid.NewGuid(), name: "Bob");
-
-        repository.Save(tenant1);
-
-        repository.Invoking(r => r.Save(tenant2))
-            .Should().Throw<DuplicateTenantException>()
-            .WithMessage($"Tenant with name `Bob` already exists.");
-    }
-
-    [Fact]
-    public void ItRejectsTwoTenantsWithTheSameGuid()
-    {
-        using var repository = new DatabaseTenantRepository(TenantContext());
-
-        var guid = Guid.NewGuid();
-        var tenant1 = GivenTenant(guid: guid, name: "Alice");
-        var tenant2 = GivenTenant(guid: guid, name: "Sussan");
-
-        repository.Save(tenant1);
-
-        repository.Invoking((r) => r.Save(tenant2))
-            .Should().Throw<DuplicateTenantException>()
-            .WithMessage($"Tenant with guid `{guid}` already exists.");
-    }
-
-    private static Tenant GivenTenant(
-        Guid? guid,
-        string name = "Stefan",
-        string email = "stefan@example.com",
-        string phone = "07925123123")
-    {
-        var theGuid = guid ?? Guid.NewGuid();
-        return new Tenant
-        {
-            Guid = theGuid,
-            Name = name,
-            ContactInfo = new Tenant.TenantContactInfo
-            {
-                Email = email,
-                Phone = phone
-            }
-        };
+        return new DatabaseTenantRepository(TenantContext());
     }
 
     private TenantContext TenantContext()
